Добавил пробелы, инициализировал переменные.
